Update TLS support
Hyper itself no longer directly supports TLS, but TLS is available (and
I don't know if this is the only option or even the best option) via a
crate by sfackler named hyper-native-tls. We now depend on that to make
encrypted calls to ICNDB.

Note that we never made any kind of encrypted call before, even if we
had the TLS feature set; the protocol used for all calls previously was
HTTP. If TLS is enabled, we now use the HTTPS protocol for calls to the
API. Whether or not this actually works, I have no idea.